requirements.txt
google-genai>=0.8.0
aiosqlite>=0.20.0
streamlit>=1.32.0
